+++ src/edit/InstanceForm.js
@@ -214,6 +214,7 @@ class InstanceForm extends React.Component {
       submitting,
       copy,
     } = this.props;
+
     const cancelButton = (
       <Button
         buttonStyle="default mega"