+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Omit `id` from preceding/succeeding titles when duplicating instance. Fixes UIIN-1324.
 * Add info callout when saving file with the instance UUIDs takes longer. Refs UIIN-1329.
 * Do not escape quotes when submitting query-search queries. Refs UIIN-1319, UIIN-1320.
+* Use `TextArea` for some data fields on instance form. Refs UIIN-1278.
 
 ## [5.0.1](https://github.com/folio-org/ui-inventory/tree/v5.0.1) (2020-10-15)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v5.0.0...v5.0.1)